server: Start server immediately
Currently, the server starts only if the OIDC setup has been completed.
That is, the oidc-authservice has contacted the well-known endpoint of
the OIDC Provider.

However, if we serve the client and provider behind the same origin (as
we do in Kubeflow) and use a public address for Dex, then the request to
the well-known endpoint exits and reenters the cluster. When it
reenters, the Istio Gateway tries to authenticate it with the
oidc-authservice.  However, the server isn't started yet and the request
to the authservice times out, resulting in the request being rejected.

This deadlock can be broken if we start the oidc-authservice server
immediately for serving whitelisted paths, like the paths to an OIDC
Provider.

Signed-off-by: Yannis Zarkadas <yanniszark@arrikto.com>
Reviewed-by: Alex Pyrgiotis <apyrgio@arrikto.com>