web-server: Add a web server for AuthService's pages
The AuthService is a server that implements Envoy's ext_authz filter.
https://www.envoyproxy.io/docs/envoy/latest/intro/arch_overview/security/ext_authz_filter#arch-overview-ext-authz

In a nutshell, the AuthService exposes a server on which Envoy will proxy
requests (headers only) to ask if they are allowed or not. A response of 200 OK
means the user's request will be proxied upstream. Any other response is sent to
the user as is.

Up until now, we have implemented the AuthService's API endpoints (oidc callback,
logout) on this server. These endpoints work because the return other-than-200
responses (they redirect), so the response is sent to the user. However, now we
want to serve some HTML pages (homepage, after_logout page) and we can't do it
on the same server as we can't send a 200 OK response (Envoy would proxy the
request and not return our response).

This made us realize that we are mixing up two different interfaces, the one
facing Envoy and the one facing the end-user. This is why, for our HTML pages
(and for all user-accessible endpoints) we introduce a web server, living
alongside the main AuthService server. Right now, the web-server only includes
the HTML pages, but the plan is to get all user-accessible endpoints under it.

The web server includes a homepage, after_logout page and theme endpoints. In
addition, we have added a document explaining how a user can create their own
templates and themes.

The web server listens at port 8082 by default.

Signed-off-by: Yannis Zarkadas <yanniszark@arrikto.com>