logout: Revoke refresh/access tokens
When the user logs out, the authservice wipes the session and the
associated data from its database (idtoken, access token, refresh
token).
However , the access/refresh token is not revoked so theoretically it
could still be used if someone managed to get a hold of it.
It would be best if we could make sure that the access/refresh token is
revoked.

OAuth2 token revocation is described by RFC7009:
https://www.rfc-editor.org/rfc/rfc7009.html

In short, a POST request is made to a revocation_endpoint.
The client MUST authenticate itself to request a token revocation.

Signed-off-by: Yannis Zarkadas <yanniszark@arrikto.com>