settings: Refactor AuthService settings
Refactor the AuthService's settings to make it easier to configure. The goal is
to have uniformly defined settings that make sense, allow deep configurability
while providing sensible defaults that cover most cases.

More specifically, we introduce a `AUTHSERVICE_URL_PREFIX` setting, which is the
URL prefix under which all AuthService user-accessible endpoints live.
Currently, those are:
- OIDC Callback URL (`/oidc/callback`)
- Logout URL (`/logout`)
- Web Pages (`/site/{homepage, after_logout, theme/*}`)

The `AUTHSERVICE_URL_PREFIX` setting allows us to provide sensible defaults for
all of the aforementioned endpoints, while keeping necessary configuration to a
minimum and retaining extra configurability if the user needs it.

In addition, we streamline setting names and format (e.g., use comma separated
lists everywhere) and refactor the README in order to present settings in an
easy-to-read matrix.

Signed-off-by: Yannis Zarkadas <yanniszark@arrikto.com>
Reviewed-by: Dimitris Aragiorgis <dimara@arrikto.com>