e2e: Add login E2E test
Add an E2E test of a complete login using the OIDC Authorization Code flow.
The test will:
1. Create a local Kubernetes cluster using k3d (https://k3d.io/).
2. Apply YAML for Dex, AuthService, Istio and a protected workload.
3. Test a complete login flow.

This test should provide an easy and quick way to validate code changes to the
AuthService and increase development velocity.

Since k3d is super lightweight, this is a very fast and cheap test even though
it uses a full-blown Kubernetes cluster.

Signed-off-by: Yannis Zarkadas <yanniszark@arrikto.com>
Reviewed-by: Dimitris Aragiorgis <dimara@arrikto.com>