logout: Change logout endpoint to return JSON response
Our logout endpoint is meant to be used by the browser's javascript, because it
demands authentication with headers. However, sending a redirect is not good,
because the XMLHTTPRequest API transparently follows redirects and doesn't allow
the javascript client to get the redirect location and set the browser window to
it.

To mitigate that, we now return a JSON response with status 201 CREATED. The
JSON response includes the redirect URL, which the javascript client will parse
and set the browser window to its value.

Signed-off-by: Yannis Zarkadas <yanniszark@arrikto.com>