settings: Don't use defaults if user has set value
Some of our settings default values depend on other settings. For those, make
sure that we check if the user has given a value before overriding it.
Update unit tests to check that scenario.

Make the default scopes openid and email, as we use email for the default claim
value.

Fix a typo in the README.

Github-PR: #26

Signed-off-by: Yannis Zarkadas <yanniszark@arrikto.com>