server: Change session cookie SameSite policy to Lax
Using a strict SameSite policy is sometimes too restrictive. For
example, an application that is protected by the AuthService may be
acting as an OIDC/OAuth client to a third-party application. In that
case, the SameSite=strict policy prevents the Provider->Client auth code
redirect from working, as it originates from the Provider origin.
Change SameSite policy to Lax, which allows cookies only from cross-site
requests / redirects using the GET method.

Signed-off-by: Yannis Zarkadas <yanniszark@arrikto.com>