Add an id token authenticator
The id token authenticator verifies the token in the header.
The header name of id token is specified by the `ID_TOKEN_HEADER`
config. The authenticator retrieves the USERID_CLAIM after the
verification. If there is no id token header, the verification failes
or the USERID_CLAIM does not exist, the request is passed to the
next authenticator.

Signed-off-by: Yihong Wang <yh.wang@ibm.com>
Reviewed-by: Yannis Zarkadas <yanniszark@arrikto.com>