Add `SESSION_SAME_SITE` option
Using a strict SameSite policy for the session cookie is sometimes too
restrictive. For example, an application that is protected by the
AuthService may be acting as an OIDC/OAuth client to a third-party
application. In that case, the SameSite=strict policy prevents the
Provider->Client auth code redirect from working, as it originates
from the Provider origin.

Make the SameSite policy of the session cookie configurable and default
to Lax, which allows cookies only from cross-site requests/redirects
using the GET method.

Signed-off-by: Yihong Wang <yh.wang@ibm.com>
Reviewed-by: Yannis Zarkadas <yanniszark@arrikto.com>