server: Implement groups authorizer
Add the notion of authorizers to the AuthService and implement an
authorizer for groups. At first, the authorizer will just take an
allowlist and use it to determine if the request is allowed or not.

In addition, remove the USERID_TOKEN_HEADER setting. With the
rearchitecture into authenticators, this functionality broke and we have
no easy way of reinstating it, as authenticators currently can't set
response header values. In the future, we may reinstate it if there is a
need OR come up with another solution (like generating our own JWTs
since we know the identity).

Github-PR: #40

Signed-off-by: Yannis Zarkadas <yanniszark@arrikto.com>
Reviewed-by: Alex Pyrgiotis <apyrgio@arrikto.com>