server: Improve UX when user is not authorized
When the user is using a browser and is not authorized, they will simply
get a 403 page with a message explaining they can't access the upstream.
However, there is no way to continue from there (e.g., logout), which
makes for a bad UX.

Remedy this by:
- Revoking the user's session if they are unauthorized.
- Not creating a session if the user is not authorized.
- Returning a message with a link to continue to login.

Refactor E2E tests accordingly to the new behavior.

In the future, the 403 will prettified and moved to the web server.

Signed-off-by: Yannis Zarkadas <yanniszark@arrikto.com>