session-authenticator: Ensure backwards-compatibility
Backwards compatibility is when older clients can make requests to newer
versions of a server and still get back correct replies. In the context
of the AuthService, the client is the AuthService in the past and the
server is the AuthService in the future. Messages are passed through the
sessions database.

In a previous commit, we introduced code that assumed a new value would
be set in sessions. This is essentially equivalent to adding a new
required field, which breaks backwards compatibility. Default the new
field to ensure backwards compatibility.

In the future, we should look into something like protocol-buffers as
the serialization format, because they offer schema evolution out of the
box.

Github-PR: #42

Signed-off-by: Yannis Zarkadas <yanniszark@arrikto.com>
Reviewed-by: Alex Pyrgiotis <apyrgio@arrikto.com>