sessions: Refactor code around sessions
In AuthService, sessions are used to track the user's logged-in status
at the AuthService, when logging in with OIDC. The session handling code
has been included inside the separate handlers so far, but now we need
to reuse some functionality. This is a good opportunity to refactor the
code around sessions and create a cleaner interface to work with them,
as well as consolidate current implementations.

Github-PR: #42

Signed-off-by: Yannis Zarkadas <yanniszark@arrikto.com>
Reviewed-by: Alex Pyrgiotis <apyrgio@arrikto.com>