security: Don't create predictable session IDs.
The code was creating nonces to use as a session ID using math/rand,
which follows a predictable sequence.  As a result an attacker could
easily predict session IDs, especially if they have the ability to
generate a session ID for themselves, then they can easily guess where
the sequence is at and trivially hijack other sessions.

This is a high severity security vulnerability.

Signed-off-by: Benoit Sigoure <tsunanet@gmail.com>