Remove CORS middleware
The AuthService used a CORS middleware as a remnant of the original
fork:
https://github.com/ajmyyra/ambassador-auth-oidc/commit/43dd5ae57de2e4bdfc786fc35dd754ffa508e357

The CORS middleware permits requests with certain default methods and
headers. However, since the default answer is 200, what it actually does
is proxy the CORS requests for those methods. We haven't yet decided how
AuthService will work with CORS and this code seems like a security risk
(e.g., what if the default response code changes tomorrow?).

As such, we remove the code until we tackle the issue of
AuthService + CORS.

Signed-off-by: Yannis Zarkadas <yanniszark@arrikto.com>