oidc: Set Path for csrf cookie
PR #51 introduced the oidc_state_csrf cookie for protecting against CSRF
attacks in the OIDC flow. However, the cookie's path is not set, meaning
that the browser will only send it for the path where it was set [1].
Thus, login cannot be completed successfully. This commit sets the
cookie's path to `/`, same as the session cookie.

[1]: https://stackoverflow.com/a/43336097

Github-PR: #57
Signed-off-by: Yannis Zarkadas <yanniszark@arrikto.com>
Reviewed-by: Alex Pyrgiotis <apyrgio@arrikto.com>