server: Add userid transformations mechanism.
There are some cases where the OIDC provider is not returning
the desired userid or that the OIDC userid must be mapped into
an internal value.

For that reason we introduce a mechanism to transform the userid that
is set in the header using a set of predefined rules that match a
regular expression and replaces it with a constant value.
The server will load the set of rules using the USERID_TRANSFORMERS
settings
in the JSON format: [{"matches": "regex", "replaces": "value"}, ...].

Github-PR: #65

Signed-off-by: Ioannis Bouloumpasis <buluba@arrikto.com>