server: Propagate the original URL
When `AFTER_LOGIN_URL` is set, after the user is authenticated
`authservice` will redirect the user to the predefined url
defined by `AFTER_LOGIN_URL` and the original path is lost.
With this commit we introduce a new query parameter `next` that
is added when redirecting to `AFTER_LOGIN_URL` that has the
value of the original url that the user first visited.

GitHub-PR: #66

Closes #67

Signed-off-by: Ioannis Bouloumpasis <buluba@arrikto.com>
Reviewed-by: Yannis Zarkadas <yanniszark@arrikto.com>