e2e: Upgrade K3d and K3s versions, fix CI breakage
1. Kernel versions after 5.12.2 made sysctl variable nf_conntrack_max
read-only for non-init net namespaces (commit 2671fa4). This change was
backported to earlier kernel versions as well.

2. K3s minor versions before {1.18.19, 1.19.11, 1.20.7, 1.21.1} by
default insist on setting this variable and break otherwise.

3. On their 2021/07/26 update, GH Actions bumped the kernel version of
the GH-hosted runners to 5.4.0-1055-azure. This kernel upgrade broke our
GH Workflow because it included change (1).

In order to fix this, we must use a K3s image that contains the fix.
Given this necessity, we seize the opportunity to also upgrade K3d to
the latest version (4.4.8) and switch from K3s minor 1.18 to 1.19, using
the latest (1.19.15) version. We also upgrade the versions of kustomize
and kubectl.

Summary of changes:
* bump K3d to 4.4.8
* bump K3s to 1.19.15
* bump kustomize to 4.4.0
* bump kubectl to 1.19.15

Signed-off-by: George Alexopoulos <grgalex@arrikto.com>
Reviewed-by: Alex Pyrgiotis <apyrgio@arrikto.com>