Narrow dependencies on k8s libraries
This change reduces the spread of the dependencies on the k8s
authenticator and k8s user authentication libraries which will make
refactoring easier.

In the former case we introduce an 'authenticator' interface and an
'authenticateResponse' interface in order to capture the existing
appearance of the k8s authenticator without declaring our use of the
implementation outside of the authenticator_kubernetes.go
implementation.

In the latter case we introduce a 'UserInfo' interface which overlaps
with the k8s userinfo interface, although we still use the latter
internally to the authenticator implementations.