jwt-authenticator: Add JWT token authenticator
Introduce a new JWT access token authenticator. This authenticator
authenticates locally the retrieved JWT access token based on the
signature of the JWT. If the `kid` is found in the list of the
stored JWKs of AuthService and the JWT access token is not expired,
then AuthService successfully authenticates the client that is
making the request.

GitHub-PR: #89

Signed-off-by: Athanasios Markou <athamark@arrikto.com>
Reviewed-by: Ioannis Bouloumpasis <buluba@arrikto.com>