server: register oauth callback handler according to REDIRECT_URL
Register callback function to the config.RedirectURL so that it works
with the REDIRECT_URL environment variable. If the REDIRECT_URL envvar
is not set by the user, then it defaults to the
<AUTHSERVICE_URL_PREFIX>/oidc/callback path.

This fix ensures that setting the REDIRECT_URL env will no longer
break the authentication flow.

GitHub-PR: #99

Signed-off-by: Junlin Zhou <jameszhou2108@hotmail.com>
Reviewed-by: Athanasios Markou <athamark@arrikto.com>