authorizers: Add new external authorizer
Introduce a new authorizer that relies on an external service to
authorize the requests.

Authservice will send a POST request to the external authorizer with
body in the following format:
{
  "timestamp": "timestamp of the request",
  "user": {
    "name": "username",
    "id": "user id",
    "groups": ["a list of groups"],
    "extra": {// object with extra info},
    "claims": {// object with the JWT claims, if JWT was used}
  },
  "request": {
    "host": "the request host",
    "method": "the request HTTP method",
    "path": "the request patch",
    "port": "the request port"
  }
}

The external service should respond with code [200, 300) if the request
is authorized or 401|403 not.

The external authorizer is enabled by setting the 'EXTERNAL_AUTHZ_URL'
to the target authorization service (e.g.
'EXTERNAL_AUTHZ_URL=http://authorizer/auth').

GitHub-PR: #98

Signed-off-by: Ioannis Bouloumpasis <buluba@arrikto.com>
Reviewed-by: Athanasios Markou <athamark@arrikto.com>