authenticators: Fix panic when Kubernetes config is missing
During initialization, authservice tries to load the Kubernetes config
for use with the Kubernetes authenticator. This prevents authservice
from running as a simple Docker container without Kubernetes.

Make loading the Kubernetes config optional if the Kuberentes
authenticator is disabled to allow the image to run as a simple
container.

Github-PR: #102

Signed-off-by: Ioannis Bouloumpasis <buluba@arrikto.com>
Reviewed-by: Athanasios Markou <athamark@arrikto.com>