server: factor-out the authentication check
Factor out the iteration over the authenticators. Introduce a new
tryAuthenticators() function. If one of the authenticators succeeds
then it will return the userInfo for the authenticated user and no
other authenticator will be examined. If no, then this value will
return nil and AuthService will initiate the Authorization Code Flow.

GitHub-PR: #

Signed-off-by: Athanasios Markou <athamark@arrikto.com>