fixes: fix discrepancies in our code
Make the log message for starting the Judje Server more clear. It
should be clear that at the point of the log message, AuthService
starts the Judge server.
Correct the README description for the SERVER_PORT configuration, its
exposed default value should be 8080 and not 8081.
Make a new cache entry for a freshly authenticated user only when the
caching mechanism is enabled.

GitHub-PR: #

Signed-off-by: Athanasios Markou <athamark@arrikto.com>