oidc: Introduce the 'oidc' package
Clean up direct imports of 'go-oidc'. Establish a clear interface for
for what we require from 'go-oidc' and isolate it inside the 'oidc'
package. This enables to consolidate the OIDC related operations in a
single place as well as to reduce irritating collisions between
'go-oidc' and the internal 'oidc' package.

GitHub-PR: #105

Signed-off-by: Mike Kelly <mkelly@arista.com>
Signed-off-by: Athanasios Markou <athamark@arrikto.com>
Reviewed-by: Ilias Katsakioris <elikatsis@arrikto.com>